requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
pandas==2.1.4
numpy==1.26.0
faiss-cpu==1.7.4
pdfplumber==0.10.4
nltk==3.8.1
scikit-learn==1.4.0
google-generativeai==0.3.2
google-api-core==2.15.0
python-dotenv==1.0.0